+++ tests/test_httpx_sync.py
@@ -331,15 +331,10 @@ def test_httpx_mock_requests_retrieval_on_same_url_and_method(httpx_mock: HTTPXM
         client.get("http://test_url", headers={"X-TEST": "test header 1"})
         client.get("http://test_url", headers={"X-TEST": "test header 2"})
 
-    assert (
-        httpx_mock.get_request(httpx.URL("http://test_url")).headers["x-test"]
-        == "test header 1"
-    )
-    assert (
-        httpx_mock.get_request(httpx.URL("http://test_url")).headers["x-test"]
-        == "test header 2"
-    )
-    assert not httpx_mock.get_request(httpx.URL("http://test_url"))
+    requests = httpx_mock.get_requests(httpx.URL("http://test_url"))
+    assert len(requests) == 2
+    assert requests[0].headers["x-test"] == "test header 1"
+    assert requests[1].headers["x-test"] == "test header 2"
 
 
 def test_httpx_mock_requests_json_body(httpx_mock: HTTPXMock):
@@ -392,3 +387,17 @@ def test_callback_returning_response(httpx_mock: HTTPXMock):
     with httpx.Client() as client:
         response = client.get("http://test_url")
         assert response.json() == {"url": "http://test_url"}
+
+
+@pytest.mark.xfail(
+    raises=AssertionError,
+    reason="Single request cannot be returned if there is more than one matching.",
+)
+def test_httpx_mock_request_retrieval_with_more_than_one(httpx_mock: HTTPXMock):
+    httpx_mock.add_response("http://test_url")
+
+    with httpx.Client() as client:
+        client.get("http://test_url", headers={"X-TEST": "test header 1"})
+        client.get("http://test_url", headers={"X-TEST": "test header 2"})
+
+    httpx_mock.get_request(httpx.URL("http://test_url"))